BottomCellItem now accepts a OnclickListener function
This allows us to directly assign it to the itemView instead of always assigning a listener to all itemViews and also, turn off / on the isClickable property, which is used by Talkback to determine whether the "Double tap to activate" hint will be announced for the view.
If a contentDescription is set in the BottomCellItem, we now set it to the itemView and mark it important for accessibility, else we set it to Auto.