Parity for legacy audioManager.isWiredHeadsetOn on API26+
When checking if headset is active:

The deprecated isWiredHeadsetOn API was checking for wired headset, wired headphone and usb headset.
For 23+ we were only checking for wired headset/headphones.

This patch unnests the logic a bit and adds a check for usb headset on API26+ where that type was introduced.