Fix crash after restart due to permission change
The Activity is attempting to resume, but since the Process is destroyed
(happens when the user turns off a permission), it can't resume the
call.

The code does attempt to gracefully shut down here (really our only
option), but it was catching the wrong exception type, and bubbling
through leading to crash.

Expected is that you can't resume after process death, but it also
shouldn't "crash", it should quiet quit back to the contoso application
(and attempt to restore the parent activity on the new process)